feat: add base auth support [ci:force]
