+++ lib/core/src/lib/auth/redirect-auth.service.ts
@@ -16,9 +16,9 @@
  */
 
 import { Inject, Injectable } from '@angular/core';
-import { AuthConfig, AUTH_CONFIG, OAuthErrorEvent, OAuthService, OAuthStorage } from 'angular-oauth2-oidc';
+import { AuthConfig, AUTH_CONFIG, OAuthErrorEvent, OAuthService, OAuthStorage, TokenResponse } from 'angular-oauth2-oidc';
 import { JwksValidationHandler } from 'angular-oauth2-oidc-jwks';
-import { Observable } from 'rxjs';
+import { from, Observable } from 'rxjs';
 import { distinctUntilChanged, filter, map, shareReplay, startWith } from 'rxjs/operators';
 import { AuthService } from './auth.service';
 
@@ -106,6 +106,20 @@ export class RedirectAuthService extends AuthService {
 
     // initLoginFlow will initialize the login flow in either code or implicit depending on the configuration
     this.ensureDiscoveryDocument().then(() => void this.oauthService.initLoginFlow(stateKey));
+  }
+
+  baseAuthLogin(username: string, password: string): Observable<TokenResponse> {
+    this.oauthService.useHttpBasicAuth = true;
+
+    return from(this.oauthService.fetchTokenUsingPasswordFlow(username, password)).pipe(
+        map((response) => {
+            const props = new Map<string, string>();
+            props.set('id_token', response.id_token);
+            // for backward compatibility we need to set the response in our storage
+            this.oauthService['storeAccessTokenResponse'](response.access_token, response.refresh_token, response.expires_in, response.scope, props);
+            return response;
+        })
+    );
   }
 
   async loginCallback(): Promise<string | undefined> {
